+++ worker/provisioner/imagemetadata.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package common
+package provisioner
 
 import (
 	"github.com/juju/errors"
